xfree86: {common,loader}: introduce option IgnoreABI for OutputClass
It is patch 1/2 of a series that provides a convenient way to specify
IgnoreABI on a per-driver basis.

Two Bool variables LoaderIgnoreABI (specific) and  LoaderIgnoresllABI
(server-wide) are used to control loader and modules behavior.

This patch also allows to create an OutputClass that applies IgnoreABI option
to devices depending on the driver in use.

Signed-off-by: Oleh Nykyforchyn <oleh.nyk@gmail.com>